use own class, use toggler
